switch to size_t and more paranoid compiler flags

+++ src/js0n.h
@@ -5,4 +5,4 @@
 // vlen = where to store return value length
 // returns pointer to value and sets len to value length, or 0 if not found
 // any parse error will set vlen to the position of the error
-char *js0n(char *key, int klen, char *json, int jlen, int *vlen);
+char *js0n(char *key, size_t klen, char *json, size_t jlen, size_t *vlen);
